migrate quickbar provider types
